Added support for cube maps,
The "visible" attribute is now properly supported for groups, meshes and lights. Some problems
remain when setting the attribute for the first time on a node that previously didn't have it defined
(eg. in the shader_lights conformance test)

Bug fix for shaders when no light source is present in the scene.

SGL Texture objects will now call XML3DHandler.redraw automatically when they're finished loading.

Added a "disablepicking" attribute to the xml3d node to improve performance in scenes that don't need it,
mouse events on the canvas itself will still be handled either way.