Revamping notification system between adapters and the renderer, part 1
All communication between adapters is done through notifyChanged now. A new event
type org.xml3d.webgl.Events.InternalMutationEvent is created after an adapter receives
a DOM event and is sent to any child adapters (in the case of group nodes). Once
it reaches a leaf (eg. mesh, light) the event is sent to the renderer which applies the change
to the drawable object itself.