renderer-utils: fixed bug in XML3D.webgl.adjustMinMax()
In that method transformation of the bbox is done based on the min-max values.

The bounding box is axis-aligned, but through transformation
min and max values might be shuffled (image e.g. for a
rotation (0, 1, 0, 1.57), the min's and max' x and z values are swapped).
So we have to order them