#30: Added bounding box interface methods for both local and world space, deprecated getBoundingBox
The interface is now divided into getLocalBoundingBox and getWorldBoundingBox. getBoundingBox is currently
retained and tied to getWorldBoundingBox but will generate a warning about being deprecated. This should make
the system more consistent with the get*Matrix interface methods and make it clear which coordinate system the
resulting BBox is in.