Add null checks in asset invalidation to prevent the RenderNodes from being created before the model has been initialized
This would happen if the user sets the "src" attribute before appending the model tag to the DOM, for example. The ModelRenderAdapter
constructor would read the visibility attribute for the node, which would trigger a FlushDOMChanges, which would
cause the model render nodes to be re-created, even though we're still in the initialization phase of the root
RenderNode for the model. This results in the render nodes being created twice and also incorrect visibility handling.