Removed Graph class and moved WebCL context to a global location (XML3D.webcl.cl).
The old way of creating the WebCL context was not per-xml3d element anyway. Due to the way the XML3DDataAdapterFactory
was created there was only ever one per document, and thus only one Graph per document. This effectively created a single
global WebCL context per document. Instead of carrying all this Graph baggage along, which wasn't used for anything except
storing the WebCL context, I've moved the context to a global variable instead.

If it really does need to be per-xml3d element we should revisit this later and design a better initialization system for it.