Adjust internal document caching to prevent multiple requests for the same document, extend request options with an allowCaching field
Relying on the browser's own document cache turned out to be a bad idea. Scenes with hundereds of requests to the same document
take up to a minute to load. Instead we now check our own document cache before sending the request out. Since this can interfere with
dynamic URLs where no-cache headers are set we also added the allowCaching field to the onRequest options so
users can disable our internal cache for specific request URLs.